requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.4
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
flask-talisman==1.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongomock==4.3.0
Naked==0.1.32
ordered-set==4.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
sentinels==1.1.1
shellescape==3.8.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
